feat: fix the esl FormESL
Added a component to deal with the image in base64, compresses the image
if necessary.
Add svg file for button in base64 component
Fix the fetch method with JSON.parse.
Changed test page to test the RoomWatcher.
Header z-index css changed so button doesnt overlap
Added masks for the inputs in FormESL
Added base64 elgin logo image in defines