feat: timer in waiting Room
changes were made to the Timer component, so that the minutes:seconds of the
timer show or not are decided by props, also took out responsability for
styling.
Changed the RoomWatcher component to adapt.
Added timer bar in waiting room usign a function that calculates how
many minutes till the expirationTime